Bump to ES 8.4.0 and fix compile errors

+++ dlic-dlsfls-legacy/src/main/java/com/floragunn/searchguard/enterprise/dlsfls/legacy/lucene/DlsFlsFilterLeafReader.java
@@ -977,7 +977,7 @@ class DlsFlsFilterLeafReader extends SequentialStoredFieldsLeafReader {
             return new SortedSetDocValues() {
 
                 @Override
-                public long docValueCount() {
+                public int docValueCount() {
                     return sortedSetDocValues.docValueCount();
                 }
